--- crates/sprinkles/tests/hashes.rs
@@ -1,1 +1,0 @@
-
